+++ lib/bugsnag.rb
@@ -14,16 +14,17 @@ require "bugsnag/delivery/synchronous"
 require "bugsnag/delivery/thread_queue"
 
 require "bugsnag/rack"
-require "bugsnag/railtie" if defined?(Rails::Railtie)
 
 require "bugsnag/middleware/rack_request"
 require "bugsnag/middleware/warden_user"
+require "bugsnag/middleware/clearance_user"
 require "bugsnag/middleware/callbacks"
 require "bugsnag/middleware/rails3_request"
 require "bugsnag/middleware/sidekiq"
 require "bugsnag/middleware/mailman"
 require "bugsnag/middleware/rake"
 require "bugsnag/middleware/callbacks"
+require "bugsnag/middleware/classify_error"
 
 module Bugsnag
   LOG_PREFIX = "** [Bugsnag] "
@@ -43,6 +44,17 @@ module Bugsnag
       # Use resque for asynchronous notification if required
       require "bugsnag/delay/resque" if configuration.delay_with_resque && defined?(Resque)
 
+      # Add info error classifier to internal middleware
+      configuration.internal_middleware.use(Bugsnag::Middleware::ClassifyError)
+
+      # Warn if an api_key hasn't been set
+      @key_warning = false unless defined?(@key_warning)
+
+      if !configuration.api_key && !@key_warning
+        warn "No API key has been set, check your configuration"
+        @key_warning = true
+      end
+
       # Log that we are ready to rock
       @logged_ready = false unless defined?(@logged_ready)
 
@@ -56,17 +68,18 @@ module Bugsnag
     def notify(exception, overrides=nil, request_data=nil, &block)
       notification = Notification.new(exception, configuration, overrides, request_data)
 
-      yield(notification) if block_given?
-
-      notification.deliver
-      notification
-    end
-
-    # Notify of an exception unless it should be ignored
-    def notify_or_ignore(exception, overrides=nil, request_data=nil, &block)
-      notification = Notification.new(exception, configuration, overrides, request_data)
+      initial_severity = notification.severity
+      initial_reason = notification.severity_reason
 
       yield(notification) if block_given?
+
+      if notification.severity != initial_severity
+        notification.severity_reason = {
+          :type => Bugsnag::Notification::USER_CALLBACK_SET_SEVERITY
+        }
+      else
+        notification.severity_reason = initial_reason
+      end
 
       unless notification.ignore?
         notification.deliver
@@ -75,13 +88,15 @@ module Bugsnag
         false
       end
     end
+    alias_method :notify_or_ignore, :notify
 
     # Auto notify of an exception, called from rails and rack exception
     # rescuers, unless auto notification is disabled, or we should ignore this
     # error class
     def auto_notify(exception, overrides=nil, request_data=nil, &block)
       overrides ||= {}
-      overrides.merge!({:severity => "error"})
+      overrides[:severity] = "error" unless overrides.has_key? :severity
+      overrides[:unhandled] = true unless overrides.has_key? :unhandled
       notify_or_ignore(exception, overrides, request_data, &block) if configuration.auto_notify
     end
 
@@ -130,7 +145,8 @@ module Bugsnag
   end
 end
 
-[:resque, :sidekiq, :mailman, :delayed_job].each do |integration|
+require "bugsnag/railtie" if defined?(Rails::Railtie)
+[:resque, :sidekiq, :mailman, :delayed_job, :shoryuken, :que].each do |integration|
   begin
     require "bugsnag/#{integration}"
   rescue LoadError