+++ lib/bugsnag/configuration.rb
@@ -31,7 +31,6 @@ module Bugsnag
     attr_accessor :proxy_password
     attr_accessor :timeout
     attr_accessor :hostname
-    attr_accessor :delivery_method
     attr_writer :ignore_classes
 
     THREAD_LOCAL_NAME = "bugsnag_req_data"
@@ -44,20 +43,6 @@ module Bugsnag
       /password/i,
       /secret/i,
       "rack.request.form_vars"
-    ].freeze
-
-    DEFAULT_IGNORE_CLASSES = [
-      "AbstractController::ActionNotFound",
-      "ActionController::InvalidAuthenticityToken",
-      "ActionController::ParameterMissing",
-      "ActionController::UnknownAction",
-      "ActionController::UnknownFormat",
-      "ActionController::UnknownHttpMethod",
-      "ActiveRecord::RecordNotFound",
-      "CGI::Session::CookieStore::TamperedWithCookie",
-      "Mongoid::Errors::DocumentNotFound",
-      "SignalException",
-      "SystemExit",
     ].freeze
 
     DEFAULT_IGNORE_USER_AGENTS = [].freeze
@@ -73,11 +58,10 @@ module Bugsnag
       self.send_environment = false
       self.send_code = true
       self.params_filters = Set.new(DEFAULT_PARAMS_FILTERS)
-      self.ignore_classes = Set.new(DEFAULT_IGNORE_CLASSES)
+      self.ignore_classes = Set.new()
       self.ignore_user_agents = Set.new(DEFAULT_IGNORE_USER_AGENTS)
       self.endpoint = DEFAULT_ENDPOINT
       self.hostname = default_hostname
-      self.delivery_method = DEFAULT_DELIVERY_METHOD
       self.timeout = 15
       self.vendor_paths = [%r{vendor/}]
       self.notify_release_stages = nil
@@ -94,6 +78,30 @@ module Bugsnag
 
       self.middleware = Bugsnag::MiddlewareStack.new
       self.middleware.use Bugsnag::Middleware::Callbacks
+    end
+
+    ##
+    # Gets the delivery_method that Bugsnag will use to communicate with the
+    # notification endpoint.
+    #
+    def delivery_method
+      @delivery_method || @default_delivery_method || DEFAULT_DELIVERY_METHOD
+    end
+
+    ##
+    # Sets the delivery_method that Bugsnag will use to communicate with the
+    # notification endpoint.
+    #
+    def delivery_method=(delivery_method)
+      @delivery_method = delivery_method
+    end
+
+    ##
+    # Used to set a new default delivery method that will be used if one is not
+    # set with #delivery_method.
+    #
+    def default_delivery_method=(delivery_method)
+      @default_delivery_method = delivery_method
     end
 
     # Accept both String and Class instances as an ignored class
@@ -124,8 +132,8 @@ module Bugsnag
     private
 
     def default_hostname
-      # Don't send the hostname on Heroku
-      Socket.gethostname unless ENV["DYNO"]
+      # Send the heroku dyno name instead of hostname if available
+      ENV["DYNO"] || Socket.gethostname;
     end
   end
 end