+++ lib/bugsnag/resque.rb
@@ -26,7 +26,16 @@ module Bugsnag
     end
 
     def save
-      Bugsnag.auto_notify(exception, {:context => "#{payload['class']}@#{queue}", :payload => payload, :delivery_method => :synchronous})
+      Bugsnag.auto_notify(exception, {
+        :context => "#{payload['class']}@#{queue}",
+        :payload => payload,
+        :severity_reason => {
+          :type => Bugsnag::Notification::UNHANDLED_EXCEPTION_MIDDLEWARE,
+          :attributes => {
+            :framework => "Resque"
+          }
+        }
+      })
     end
   end
 end
@@ -37,7 +46,14 @@ Resque::Failure::Bugsnag = Bugsnag::Resque
 # Auto-load the failure backend
 Bugsnag::Resque.add_failure_backend
 
-Resque.before_first_fork do
-  Bugsnag.configuration.app_type = "resque"
-  Bugsnag.configuration.delivery_method = :synchronous
+if Resque::Worker.new(:bugsnag_fork_check).fork_per_job?
+  Resque.after_fork do
+    Bugsnag.configuration.app_type = "resque"
+    Bugsnag.configuration.default_delivery_method = :synchronous
+  end
+else
+  Resque.before_first_fork do
+    Bugsnag.configuration.app_type = "resque"
+    Bugsnag.configuration.default_delivery_method = :synchronous
+  end
 end